+++ lib/async_queue.dart
@@ -7,3 +7,8 @@
 library async_queue;
 
 export 'src/async_queue_base.dart';
+export 'src/job_info.dart';
+export 'src/typedef.dart';
+export 'src/queue_event.dart';
+export 'src/exceptions.dart';
+export 'src/async_node.dart' hide AsyncNode;