+++ lib/src/async_node.dart
@@ -1,6 +1,23 @@
+import 'job_info.dart';
 import 'typedef.dart';
 
-enum JobState { pending, running, failed, done }
+/// states of a job
+enum JobState {
+  ///pending
+  pending,
+
+  /// job is running
+  running,
+
+  ///job failed and removed from the queue
+  failed,
+
+  ///job done and removed from the queue
+  done,
+
+  ///job failed and pending retry
+  pendingRetry,
+}
 
 ///AsyncNode
 ///
@@ -8,6 +25,8 @@ enum JobState { pending, running, failed, done }
 class AsyncNode {
   final AsyncJob _job;
   final int maxRetry;
+  final String label;
+  final String? description;
 
   AsyncNode? next;
   int retryCount = 0;
@@ -15,6 +34,8 @@ class AsyncNode {
 
   AsyncNode({
     required AsyncJob job,
+    required this.label,
+    this.description,
     this.maxRetry = 1,
   }) : _job = job;
 
@@ -22,4 +43,17 @@ class AsyncNode {
     state = JobState.running;
     await _job();
   }
+
+  @override
+  String toString() {
+    return 'AsyncNode(maxRetry: $maxRetry, label: $label, description: $description, retryCount: $retryCount)';
+  }
+
+  JobInfo get info => JobInfo(
+        label: label,
+        description: description,
+        maxRetry: maxRetry,
+        retryCount: retryCount,
+        state: state,
+      );
 }