+++ test/async_queue_retry_test.dart
@@ -66,6 +66,7 @@ void main() {
 
   test('Default retry should be 1', () async {
     final q = AsyncQueue();
+    q.addQueueListener((event) => print(event));
     int jobCount = 0;
 
     q.addJob(
@@ -81,5 +82,62 @@ void main() {
 
     await q.start();
     expect(jobCount, 2);
+  });
+
+  test('Job should retry until success if retry time is -1', () async {
+    final q = AsyncQueue();
+    int jobCount = 0;
+    int retryCount = 0;
+    final failTime = 12;
+
+    q.addJob(
+        () => Future.delayed(
+              const Duration(milliseconds: 100),
+              () async {
+                try {
+                  jobCount++;
+                  if (jobCount <= failTime) {
+                    throw Exception('error');
+                  }
+                } catch (e) {
+                  retryCount++;
+                  q.retry();
+                }
+              },
+            ),
+        retryTime: -1);
+
+    await q.start();
+    print(jobCount);
+    expect(jobCount, failTime + 1);
+    expect(retryCount, failTime);
+  });
+  test('Job should not retry  if retry is not called even retryTime is -1',
+      () async {
+    final q = AsyncQueue();
+    int jobCount = 0;
+    int retryCount = 0;
+    final failTime = 12;
+
+    q.addJob(
+        () => Future.delayed(
+              const Duration(milliseconds: 100),
+              () async {
+                try {
+                  jobCount++;
+                  if (jobCount <= failTime) {
+                    // throw Exception('error');
+                  }
+                } catch (e) {
+                  retryCount++;
+                }
+              },
+            ),
+        retryTime: -1);
+
+    await q.start();
+    print(jobCount);
+    expect(jobCount, 1);
+    expect(retryCount, 0);
   });
 }